STRUCT: Continue moving cik_lookup
